minor: moved the one-project view to a dynamic route and left the main /projects page not implemented for now.
